Fix some log_Printf() casting; NetBSD (pkgsrc)

+++ usr.sbin/ppp/ppp/atm.c
@@ -23,7 +23,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	$OpenBSD: atm.c,v 1.2 2000/09/14 22:03:23 brian Exp $
+ *	$OpenBSD: atm.c,v 1.3 2001/01/29 08:47:43 brian Exp $
  */
 
 #include <sys/types.h>
@@ -79,7 +79,7 @@ atm_Sendto(struct physical *p, const void *v, size_t n)
 {
   ssize_t ret = write(p->fd, v, n);
   if (ret < 0) {
-    log_Printf(LogDEBUG, "atm_Sendto(%d): %s\n", n, strerror(errno));
+    log_Printf(LogDEBUG, "atm_Sendto(%ld): %s\n", (long)n, strerror(errno));
     return ret;
   }
   return ret;
@@ -90,7 +90,7 @@ atm_Recvfrom(struct physical *p, void *v, size_t n)
 {
     ssize_t ret = read(p->fd, (char*)v, n);
     if (ret < 0) {
-      log_Printf(LogDEBUG, "atm_Recvfrom(%d): %s\n", n, strerror(errno));
+      log_Printf(LogDEBUG, "atm_Recvfrom(%ld): %s\n", (long)n, strerror(errno));
       return ret;
     }
     return ret;
